Vendor import of clang RELEASE_360/rc1 tag r226102 (effectively, 3.6.0 RC1): https://llvm.org/svn/llvm-project/cfe/tags/RELEASE_360/rc1@226102
Former-commit-id: 06d4ba388873e6d1cfa9cd715a8935ecc8cd2097

+++ test/CodeGen/avx2-builtins.c
@@ -30,7 +30,7 @@ name|__m256i
 name|y
 parameter_list|)
 block|{
-comment|// CHECK: @llvm.x86.avx2.mpsadbw({{.*}}, {{.*}}, i32 3)
+comment|// CHECK: @llvm.x86.avx2.mpsadbw({{.*}}, {{.*}}, i8 3)
 return|return
 name|_mm256_mpsadbw_epu8
 argument_list|(
